Disable Obsolete 63 Byte Report Per Firmware Author's Advice
This change applies the fix recommended by @mlyonupdesigns here:
https://github.com/wemakewaves/PyRow/issues/5

Page 6 of the PM5 CSAFE specification indicates that the spec was revised on
04/06/23 by Mark Lyons, which is likely the change mentioned in the linked
issue:

    Changed USB report ID 4 size from 62 to 500 bytes; added some a dditonal
    enumeration definitions; V0.26

See: https://www.concept2.com/files/pdf/us/monitors/PM5_CSAFECommunicationDefinition.pdf

This change is said to be necessary for compatibility with PM5 ergs, which
happens to be what I have on hand.  I have confirmed that this commit enables
the statshow.log command to function correctly with a PM5 Concept2 RowErg.

Additional information:  I am runing this on NixOS over USB with 3.13.